requirements.txt
requests>=2.28.1
urllib3>=1.26.12
python-dateutil>=2.8.2
argparse>=1.4.0
pyinstaller>=5.3